Added edge v1/v2 cross send test
